+++ src/index.turbo.ts
@@ -12,6 +12,7 @@ import Tabs from './components/tabs';
 import Tooltip from './components/tooltip';
 import InputCounter from './components/input-counter';
 import CopyClipboard from './components/clipboard';
+import Datepicker from './components/datepicker';
 import { initFlowbite } from './components';
 import Events from './dom/events';
 
@@ -54,5 +55,6 @@ export default {
     Tooltip,
     InputCounter,
     CopyClipboard,
+    Datepicker,
     Events,
 };