+++ content/getting-started/changelog.md
@@ -15,6 +15,32 @@ nextLink: getting-started/react/
 
 We strive to keep a good accountability of all of the version changes that we make for the Flowbite library.
 
+### v2.5.2
+
+- released WYSIWYG plugin component
+
+### v2.5.1
+
+- fix plugin import error for the config file
+
+### v2.5.0
+
+- added new datatables plugin and examples
+- other minor bug fixes and improvements
+
+### v2.4.1
+
+- the datepicker is now a core component of Flowbite and has API methods, events, and options
+- updated the documentation for the datepicker component and related integration guides
+- minor visual bug fixes and improvements
+
+Note: if you're using Nuxt or Angular with SSR please check the new guides:
+
+- [Nuxt guide](https://flowbite.com/docs/getting-started/nuxt-js/#data-attributes)
+- [Angular guide](https://flowbite.com/docs/getting-started/angular/#using-with-angular-ssr)
+
+These should be the only ways you import Flowbite since we need to make sure that the client (ie. browser) is available.
+
 ### v2.3.0
 
 - added new "copy to clipboard" component and examples